requirements.txt
praw
python-dotenv
anthropic
fastapi
uvicorn[standard]
pydantic